UITEST Config, build and test project appear to be working.
