--- ClubOrganiser/ClubOrganiser.UITests/AppInitializer.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using Xamarin.UITest;
-using Xamarin.UITest.Queries;
-
-namespace ClubOrganiser.UITests
-{
-    public class AppInitializer
-    {
-        public static IApp StartApp(Platform platform)
-        {
-            if (platform == Platform.Android)
-            {
-                return ConfigureApp.Android.StartApp();
-            }
-
-            return ConfigureApp.iOS.StartApp();
-        }
-    }
-}
