--- ClubOrganiser/ClubOrganiser.UITests/Tests.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using NUnit.Framework;
-using Xamarin.UITest;
-using Xamarin.UITest.Queries;
-
-namespace ClubOrganiser.UITests
-{
-    [TestFixture(Platform.Android)]
-    [TestFixture(Platform.iOS)]
-    public class Tests
-    {
-        IApp app;
-        Platform platform;
-
-        public Tests(Platform platform)
-        {
-            this.platform = platform;
-        }
-
-        [SetUp]
-        public void BeforeEachTest()
-        {
-            app = AppInitializer.StartApp(platform);
-        }
-
-        [Test]
-        public void WelcomeTextIsDisplayed()
-        {
-            AppResult[] results = app.WaitForElement(c => c.Marked("Welcome to Xamarin.Forms!"));
-            app.Screenshot("Welcome screen.");
-
-            Assert.IsTrue(results.Any());
-        }
-    }
-}
